extract more fns
